move apple snooze view and supporting code to separate file
